Remove BuildContext
BuildContext was once a convenient way to pass a group of standard tools
into each operator as it was built. This includes a logger, arbitrary
parameters, a database file, namespace information, etc.

Recently, the struct has been stripped of its usefulness, and can now be
removed entirely in order to simplify the codebase.